backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
cffi==2.0.0
cryptography==46.0.2
Brotli==1.1.0
cffi==2.0.0
cryptography==46.0.2
cssselect2==0.8.0
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
curl_cffi==0.13.0
celery>=5.3.0
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
fonttools==4.60.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
pycparser==2.23
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.1.1
python-http-client==3.3.7
frozendict==2.4.6
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multitasking==0.0.12
numpy==2.2.6
pandas==2.3.3
peewee==3.18.2
platformdirs==4.4.0
protobuf==6.32.1
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis>=4.5.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sendgrid==6.12.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
uritemplate==4.2.0
Werkzeug==3.1.3
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.66
